Simplify splitting neighbour context
