+++ package.yaml
@@ -47,7 +47,7 @@ dependencies:
 
 library:
   source-dirs: src
-  ghc-options: -fwarn-redundant-constraints
+  ghc-options: -fwarn-redundant-constraints -Wall
 
 
 tests:
